requirements.txt
aiohttp
anthropic
arxiv   
datasets
evaluate
huggingface-hub
ipython
jiter
numpy
openai
pandas
psycopg2-binary
pytorch-lightning
redis
requests
scikit-learn
scipy
sentence-transformers
SQLAlchemy
tiktoken
torch
torchaudio
torchmetrics
torchvision
transformers
#wandb
accelerate
python-dotenv
rich